Pages are now more functional. A page can now contain sub-pages, and it can send clicks from an outer page into objects within sub-pages. The entire calculate_pages() function has been re-written to organize each module as a page within a larger page that contains multiple modules. This larger page is part of the list of pages accessible to the user.

+++ function_forwarder.cpp
@@ -60,7 +60,7 @@ void oscillator_function_forwarder(Graphics_Object *g)
     {
         Text_Box *frequency_text_box = (Text_Box *) g;
         SDL_LockAudio();
-        oscillator->audio.frequency = stoi(frequency_text_box->text->current_text.c_str());
+        oscillator->audio.frequency = stof(frequency_text_box->text->current_text.c_str());
         oscillator->audio.shifted_frequency = oscillator->audio.frequency;
         SDL_UnlockAudio();
         cout << oscillator->name << " frequency changed" << endl;
